# apps/api/swag/openapi3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vyzorix Update Server API",
    "version": "0.0.01",
    "description": "Go-generated OpenAPI contract from handler annotations"
  },
  "paths": {
...
    "/annotations": {
      "get": {
        "summary": "",
        "description": "",
        "tags": [
          "annotations",
          "annotations"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      },
      "post": {
        "summary": "",
        "description": "",
        "tags": [
          "annotations",
          "annotations"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
    "/annotations/{id}": {
      "delete": {
        "summary": "",
        "description": "",
        "tags": [
          "annotations",
          "annotations"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
...
    "/channels/status": {
      "get": {
        "summary": "",
        "description": "",
        "tags": [
          "channels"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
    "/channels/subscribe": {
      "post": {
        "summary": "",
        "description": "",
        "tags": [
          "channels"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
    "/channels/unsubscribe": {
      "post": {
        "summary": "",
        "description": "",
        "tags": [
          "channels"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
...
  },
...
}